refactor(sdk)!: move EnqueueResult to the helpers queue submodule
EnqueueResult now lives in @iii-dev/helpers/queue and is removed from the root
SDK public surface (no re-export shim). The root SDK depends on iii-helpers and
imports it internally for the trigger() return type. TriggerActionEnqueue stays
at the root SDK; Rust keeps Enqueue as a core TriggerAction variant.
