+++ sdk/packages/node/helpers/src/queue/index.ts
@@ -1,1 +1,7 @@
-export {}
+/**
+ * Result returned when a function is invoked with `TriggerAction.Enqueue`.
+ */
+export type EnqueueResult = {
+  /** Unique receipt ID for the enqueued message. */
+  messageReceiptId: string
+}
