+++ sdk/packages/node/iii-example/package.json
@@ -7,6 +7,7 @@
     "dev": "bun src/index.ts"
   },
   "dependencies": {
+    "@iii-dev/helpers": "workspace:*",
     "iii-sdk": "workspace:*",
     "zod": "^4.3.6"
   },
