+++ sdk/packages/rust/iii/src/protocol.rs
@@ -69,13 +69,6 @@ pub enum TriggerAction {
     Void,
 }
 
-/// Result returned by the engine when a message is successfully enqueued.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct EnqueueResult {
-    #[serde(rename = "messageReceiptId")]
-    pub message_receipt_id: String,
-}
-
 /// Request object for `trigger()`. Matches the Node/Python SDK signature:
 /// `trigger({ function_id, payload, action?, timeout_ms? })`
 ///
